Use localStorage to only show confetti once
